♻️ refactor: 캐러셀 이미지 webp로 변환하여 최적화
