+++ src/App.css
@@ -71,6 +71,7 @@
 .animate-slide-down {
   animation: slideDown 0.5s ease-out;
 }
+
 /* 자주 사용하는 스타일을 재사용하기 위해 컴포넌트화*/
 @layer components {
   .flex-center {
