+++ src/components/recruting/home/_main/Carousel.css
@@ -1,0 +1,13 @@
+.slick-dots {
+  bottom: 20px !important;
+  z-index: 1;
+}
+
+.slick-dots li button:before {
+  color: white !important;
+  font-size: 10px !important;
+}
+
+.slick-dots li.slick-active button:before {
+  color: white !important;
+}
